+++ src/viewer/background-preview.test.ts
@@ -46,6 +46,8 @@ function stageWith(
       far: 0,
       fov: 0,
       yShift: 0,
+      xScale: 1,
+      yScale: 1,
     },
     elements,
     animations,